test(p2p): move wire tests into the tests tree
