+++ crates/p2p/tests/wire/main.rs
@@ -1,0 +1,5 @@
+mod car;
+mod manage;
+mod manage_correlator;
+mod pubsub;
+mod se_correlator;
